+++ backend/src/controllers/persone.controller.js
@@ -49,7 +49,7 @@ function createInterna(req, res, next) {
     //recupera json
     const personaInterna = req.body;
 
-    service.createInterna(idSede, req.user.id, personaInterna, (err, results) => {
+    service.createInterna(idSede, { userId: req.user.id, username: req.user.username, ip: req.ip }, personaInterna, (err, results) => {
 
         if (err) {
             return next(err);
@@ -72,7 +72,7 @@ function updateInterna(req, res, next) {
     const idSede = req.params.idS;
     const personaInterna = req.body;
 
-    service.updateInterna(id, idSede, req.user.id, personaInterna, (err, results) => {
+    service.updateInterna(id, idSede, { userId: req.user.id, username: req.user.username, ip: req.ip }, personaInterna, (err, results) => {
 
         if (err) {
             return next(err);
@@ -102,7 +102,7 @@ function remove(req, res, next) {
 
     const id = req.params.id;
 
-    service.remove(id, req.user.id, (err, result) => {
+    service.remove(id, { userId: req.user.id, username: req.user.username, ip: req.ip }, (err, result) => {
 
         if (err) {
             return next(err);
